実装をヘッダファイルに移し、defineマクロで解禁できるようにした。 （従来のままだと、使用しない場合でもMySQL-Connectorをインストールする必要があったため。）
