+++ YanaPServer/YanaPServer/include/YanaPServer/MySQL/MySqlConnection.h
@@ -1,6 +1,8 @@
 #ifndef __MYSQLCONNECTION_H__
 #define __MYSQLCONNECTION_H__
 
+#ifdef USE_MYSQL
+
 #ifdef _WIN32
 #include <windows.h>
 #include <mysql.h>
@@ -27,12 +29,18 @@ public:
 	/**
 	 * @brief コンストラクタ
 	 */
-	CMySqlConnection();
+	CMySqlConnection()
+		: pConnection(NULL)
+	{
+	}
 
 	/**
 	 * @brief デストラクタ
 	 */
-	~CMySqlConnection();
+	~CMySqlConnection()
+	{
+		Disconnect();
+	}
 
 	/**
 	 * @fn bool Connect(const char *pHost, const char *pUserName, const char *pPassword, const char *pDBName)
@@ -43,7 +51,22 @@ public:
 	 * @param[in] pDBName ＤＢ名
 	 * @return 成功したらtrueを返す
 	 */
-	bool Connect(const char *pHost, const char *pUserName, const char *pPassword, const char *pDBName);
+	bool Connect(const char *pHost, const char *pUserName, const char *pPassword, const char *pDBName)
+	{
+		if (pConnection != NULL)
+		{
+			Disconnect();
+		}
+
+		pConnection = mysql_init(NULL);
+		if (pConnection == NULL) { return false; }
+
+		mysql_options(pConnection, MYSQL_SET_CHARSET_NAME, "utf8");
+
+		if (!mysql_real_connect(pConnection, pHost, pUserName, pPassword, pDBName, 0, NULL, 0)) { return false; }
+
+		return true;
+	}
 
 	/**
 	 * @fn bool SimpleQuery(const char *pQuery) const
@@ -51,7 +74,13 @@ public:
 	 * @param[in] pQuery クエリ
 	 * @return 成功したらtrueを返す
 	 */
-	bool SimpleQuery(const char *pQuery) const;
+	bool SimpleQuery(const char *pQuery) const
+	{
+		if (pConnection == NULL) { return false; }
+
+		int Result = mysql_query(pConnection, pQuery);
+		return (Result == 0);
+	}
 
 	/**
 	 * @fn CMySqlQuery CreateQuery(const char *pQuery) const
@@ -59,13 +88,23 @@ public:
 	 * @param[in] pQuery クエリ
 	 * @return クエリオブジェクト
 	 */
-	CMySqlQuery CreateQuery(const char *pQuery) const;
+	CMySqlQuery CreateQuery(const char *pQuery) const
+	{
+		CMySqlQuery Query(pConnection, pQuery);
+		return Query;
+	}
 
 	/**
 	 * @fn void Disconnect()
 	 * @brief 切断
 	 */
-	void Disconnect();
+	void Disconnect()
+	{
+		if (pConnection == NULL) { return; }
+
+		mysql_close(pConnection);
+		pConnection = NULL;
+	}
 
 private:
 
@@ -76,5 +115,7 @@ private:
 
 }
 }
+
+#endif		// #ifdef USE_MYSQL
 
 #endif		// #ifndef __MYSQLCONNECTION_H__
